fix(native-http-backend): send JSON primitives and multipart/form-data
* fix(native-http-backend): send JSON primitives and multipart/form-data

The original purpose of the change is to send JSON primitives with application/json requests. The feature had required a rework of serializer and data detection. As a result, the commit is also fixing a number of bugs such as supporting application/x-www-form-urlencoded requests, sending HttpParams object, and multiple values in headers.

fix #111, fix #103, fix #101, fix #64, fix #31

* refactor(detect-serializer-and-data): replace ?? with ||